TACKLE-58: moved MultipartImportBody and added RestClient annotation in tests
