+++ src/main/java/io/tackle/applicationinventory/service/ImportService.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.dataformat.csv.CsvMapper;
 import com.fasterxml.jackson.dataformat.csv.CsvSchema;
-import io.tackle.applicationimporter.MultipartImportBody;
 import io.tackle.applicationinventory.BusinessService;
+import io.tackle.applicationinventory.MultipartImportBody;
 import io.tackle.applicationinventory.TagType;
 import io.tackle.applicationinventory.entities.ApplicationImport;
 import io.tackle.applicationinventory.exceptions.ApplicationsInventoryException;
